refactor: remaned to jwkStorage and made terminateLogin a non-static function
