+++ test-integration/helpers/externalUser.ts
@@ -24,7 +24,7 @@ import { LoginRequest } from '../../src/sdk/util/request';
 import { objToBase64Url } from '../../src/sdk/util/base64';
 import { VerifiableCredential } from '../../src/sdk/util/ssi/vc';
 import { getAccount } from '../../src/sdk/services/blockchain';
-import { getJwkIssuerFromStorage } from '../../src/sdk/helpers/jwk';
+import { getJwkIssuerFromStorage } from '../../src/sdk/helpers/jwkStorage';
 import { verifyRequests } from '../../src/sdk/helpers/requests';
 import { getLoginRequestFromUrl, onRedirectLogin } from '../../src/sdk/helpers/urls';
